Migrate auth JDBC example to auth SQL
