+++ web-examples/src/main/resources/io/vertx/example/web/authsql/private/private_page.html
