+++ web-examples/src/main/resources/io/vertx/example/web/authsql/webroot/loginpage.html
