+++ web-examples/src/main/resources/io/vertx/example/web/authsql/webroot/page1.html
